refactor: refatora o  EmployeeController para incluir o ResponseEntity nas minhas rotas e customizar a resposta das requisições do cliente

+++ src/main/java/com/lorenzo/CadastroDeFuncionarios/Employees/Controller/EmployeeController.java
@@ -1,11 +1,13 @@
 package com.lorenzo.CadastroDeFuncionarios.Employees.Controller;
 
-import com.lorenzo.CadastroDeFuncionarios.Employees.Model.EmployeeModel;
 import com.lorenzo.CadastroDeFuncionarios.Employees.Service.EmployeesService;
 import com.lorenzo.CadastroDeFuncionarios.Employees.dto.EmployeeDTO;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/employees")
@@ -24,32 +26,49 @@ public class EmployeeController {
 
     // Add Employee (CREATE)
     @PostMapping
-    public EmployeeDTO addEmployee(@RequestBody EmployeeDTO dto) {
-        return employeesService.addEmployee(dto);
+    public ResponseEntity<String> addEmployee(@RequestBody EmployeeDTO dto) {
+        EmployeeDTO add = employeesService.addEmployee(dto);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Funcionario: " + add.getName() + " criado com sucesso | ID: " + add.getId() );
     }
 
     // Show all Employees (READ)
     @GetMapping
-    public List<EmployeeDTO> showAllEmployees() {
-        return employeesService.showAllEmployees();
+    public ResponseEntity<List<EmployeeDTO>> showAllEmployees() {
+        return ResponseEntity.ok(employeesService.showAllEmployees());
     }
 
     // Search Employee by ID (READ)
     @GetMapping("/{id}")
-    public EmployeeDTO showEmployeeById(@PathVariable Long id) {
-        return employeesService.showEmployeeById(id);
+    public ResponseEntity<?> showEmployeeById(@PathVariable Long id) {
+        EmployeeDTO findId = employeesService.showEmployeeById(id);
+        if (findId != null) {
+            return ResponseEntity.ok(findId);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Funcionario com ID: "+ id +" | Não encontrado");
     }
 
     // Update Employee data (UPDATE)
     @PutMapping("/{id}")
-    public EmployeeDTO updateEmployee(@RequestBody EmployeeDTO dto, @PathVariable Long id) {
-        return employeesService.updateEmployee(dto, id);
+    public ResponseEntity<String> updateEmployee(@RequestBody EmployeeDTO dto, @PathVariable Long id) {
+        if (employeesService.showEmployeeById(id) != null) {
+            employeesService.updateEmployee(dto, id);
+            return ResponseEntity.ok("Funcionario " + dto.getName() + " ID: " + id + " | Atualizado com sucesso");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Funcionario de ID: " + id + " | Não encontrado");
     }
 
     // Delete Employee (DELETE)
     @DeleteMapping("/{id}")
-    public void removeEmployee(@PathVariable Long id) {
-        employeesService.removeEmployee(id);
+    public ResponseEntity<String> removeEmployee(@PathVariable Long id) {
+        if (employeesService.showEmployeeById(id) != null) {
+            employeesService.removeEmployee(id);
+            return ResponseEntity.ok("Funcionario de id: " + id + " | Deletado com sucesso");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Funcionario de id: " + id + " | Não encontrado");
     }
 
 }
